correct the default export
